+++ packages/essdiffraction/src/ess/powder/__init__.py
@@ -14,7 +14,6 @@ from . import (
     grouping,
     masking,
     smoothing,
-    transform,
     workflow,
 )
 from .correction import RunNormalization
@@ -50,7 +49,6 @@ __all__ = [
     "masking",
     "providers",
     "smoothing",
-    "transform",
     "with_pixel_mask_filenames",
     "workflow",
 ]
